feat(calendar): move google calendar connection into app hub panel
The Google Calendar connect and disconnect flow was buried inside the
Account settings tab making it hard to find next to profile and password
fields. Users expected it alongside the other integrations in the Apps hub.

Moved the full OAuth connect/disconnect flow (launchWebAuthFlow, code
exchange, error handling) into CalendarSettings which is rendered inside
the Calendar app panel. If the user is not signed in the panel shows a
notice with a link to Account. When signed in it shows the integration
card with connect/disconnect and the look-ahead window dropdown only when
connected. Removed the Google Calendar card, IntegrationCard component,
and all related helpers from AccountSettings.
